Exercise 1/requirements.txt
pyspark==3.5.1
kafka-python-ng==2.2.2
pandas
